Split gif mod.rs into to_proto.rs.
